refactor: separate dm component
